# docs/swagger.yaml
basePath: /
definitions:
...
  commons.Pagination:
    properties:
      page_number:
        type: integer
      page_size:
        type: integer
      total_item:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.CreateInvoiceRequest:
    properties:
      customer_id:
        type: string
      due_date:
        type: integer
      invoice_date:
        type: integer
      invoice_number:
        type: string
      items:
        items:
          $ref: '#/definitions/valueobject.InvoiceItem'
        type: array
      notes:
        type: string
      quote_id:
        type: string
    type: object
...
  models.Customer:
    properties:
      address:
        type: string
      country:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: string
      metadata:
        $ref: '#/definitions/valueobject.Metadata'
      name:
        type: string
      phone:
        type: string
      picture_url:
        type: string
      postal_code:
        type: string
      social_name:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.Invoice:
    properties:
      customer:
        $ref: '#/definitions/models.Customer'
      due_date:
        type: integer
      invoice_date:
        type: integer
      invoice_id:
        type: string
      invoice_number:
        type: string
      items:
        items:
          $ref: '#/definitions/valueobject.InvoiceItem'
        type: array
      metadata:
        $ref: '#/definitions/valueobject.Metadata'
      notes:
        type: string
      payment_date:
        type: integer
      payment_method:
        type: string
      payment_status:
        $ref: '#/definitions/valueobject.PaymentStatus'
      quote_id:
        type: string
    type: object
  models.Invoices:
    properties:
      invoices:
        items:
          $ref: '#/definitions/models.Invoice'
        type: array
      pagination:
        $ref: '#/definitions/commons.Pagination'
    type: object
...
  models.UpdateInvoiceRequest:
    properties:
      customer_id:
        type: string
      due_date:
        type: integer
      invoice_date:
        type: integer
      invoice_number:
        type: string
      items:
        items:
          $ref: '#/definitions/valueobject.InvoiceItem'
        type: array
      notes:
        type: string
      payment_status:
        $ref: '#/definitions/valueobject.PaymentStatus'
      quote_id:
        type: string
    type: object
...
  valueobject.InvoiceItem:
    properties:
      description:
        type: string
      quantity:
        type: integer
      tax_rate:
        type: number
      unit_price:
        type: number
    type: object
  valueobject.Metadata:
    properties:
      created_at:
        type: integer
      created_by:
        type: string
      deleted_at:
        type: integer
      deleted_by:
        type: string
      updated_at:
        type: integer
      updated_by:
        type: string
    type: object
  valueobject.PaymentStatus:
    enum:
    - PAYMENT_STATUS_UNSPECIFIED
    - UNPAID
    - PAID
    - OVERDUE
    type: string
    x-enum-varnames:
    - PaymentStatusUnspecified
    - PaymentStatusUnpaid
    - PaymentStatusPaid
    - PaymentStatusOverdue
...
host: localhost:9000
info:
  contact:
    email: orkys.com@gmail.com
    name: API Support
    url: http://www.example.com/support
  description: This is the API documentation for the Orkys API Gateway
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: Orkys API
  version: "1.0"
paths:
...
  /invoice:
    get:
      consumes:
      - application/json
      description: Returns a list of invoices with pagination.
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of invoices per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: The invoice list and pagination details
          schema:
            $ref: '#/definitions/models.Invoices'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Returns a list of invoices.
      tags:
      - invoice
    post:
      consumes:
      - application/json
      description: Creates a new invoice with the provided details.
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: The invoice object
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/models.CreateInvoiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: The created invoice object
          schema:
            $ref: '#/definitions/models.Invoice'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Creates a new invoice.
      tags:
      - invoice
  /invoice/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an invoice by ID.
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Deletes an invoice.
      tags:
      - invoice
    get:
      consumes:
      - application/json
      description: Returns an invoice by ID with all the details.
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The invoice object
          schema:
            $ref: '#/definitions/models.Invoice'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Returns an invoice by ID.
      tags:
      - invoice
    patch:
      consumes:
      - application/json
      description: Updates an invoice with the provided details.
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: The invoice object
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/models.UpdateInvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: The updated invoice object
          schema:
            $ref: '#/definitions/models.Invoice'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Updates an invoice.
      tags:
      - invoice
...
swagger: "2.0"